+++ src/handler/handlerUpload.js
@@ -1,4 +1,4 @@
-import path from 'node:path';
+//import path from 'node:path';
 import multer from 'multer';
 
 /**
@@ -17,11 +17,13 @@ import multer from 'multer';
 export const handlerUpload = () => {
 	const upload = multer({
 		storage: multer.diskStorage({
-			destination: '/tmp/uploads',
+			/*
+			destination: '/tmp/webplsql',
 			filename: (req, file, cb) => {
 				const uniqueSuffix = `${Date.now()}-${Math.round(Math.random() * 1e9)}`;
 				cb(null, `${file.fieldname}-${uniqueSuffix}${path.extname(file.originalname)}`);
 			},
+			*/
 		}),
 		limits: {
 			fileSize: 50 * 1024 * 1024, // 50MB limit