+++ src/types.js
@@ -3,6 +3,7 @@ import z from 'zod';
 /**
  * @typedef {import('oracledb').BindParameter} BindParameter
  * @typedef {import('express').CookieOptions} CookieOptions
+ * @typedef {import('express').Request} Request
  */
 
 /**
@@ -21,6 +22,17 @@ export const z$configStaticType = z.strictObject({
 });
 
 /**
+ * @callback transactionCallbackType
+ * @param {Request} req - Incoming request object.
+ * @param {import('oracledb').Connection} connection - Active database connection pool.
+ * @returns {void | Promise<void>}
+ */
+
+/**
+ * @typedef {'commit' | 'rollback' | transactionCallbackType | undefined | null} transactionModeType
+ */
+
+/**
  * @typedef {object} configPlSqlHandlerType
  * @property {string} defaultPage - The default page.
  * @property {string} [pathAlias] - The path alias.
@@ -29,6 +41,10 @@ export const z$configStaticType = z.strictObject({
  * @property {string[]} [exclusionList] - The exclusion list.
  * @property {string} [requestValidationFunction] - The request validation function.
  * @property {Record<string, string>} [cgi] - The additional CGI.
+ * @property {transactionModeType} [transactionMode='commit'] - Specifies an optional transaction mode.
+ * "commit" this automatically commits any open transaction after each request. This is the defaults because this is what mod_plsql and ohs are doing.
+ * "rollback" this automatically rolles back any open transaction after each request.
+ * "transactionCallbackType" this allows to defined a custom handler as a JavaScript function.
  * @property {errorStyleType} errorStyle - The error style.
  */
 export const z$configPlSqlHandlerType = z.strictObject({
@@ -38,6 +54,7 @@ export const z$configPlSqlHandlerType = z.strictObject({
 	documentTable: z.string(),
 	exclusionList: z.array(z.string()).optional(),
 	requestValidationFunction: z.string().optional(),
+	transactionMode: z.unknown().optional(),
 	errorStyle: z$errorStyleType,
 });